re-remove zlib submodule

+++ inc/types.hxx
@@ -8,8 +8,8 @@ namespace types
     typedef unsigned short u16;
     typedef signed short s16;
 
-    typedef unsigned long u32;
-    typedef signed long s32;
+    typedef unsigned int u32;
+    typedef signed int s32;
 
     typedef unsigned long long u64;
     typedef signed long long s64;
